+++ crates/server/src/api/routes/quests/mod.rs
@@ -2,7 +2,7 @@ pub mod activate_quest;
 pub mod create_quest;
 pub mod delete_quest;
 pub mod get_quest;
-pub mod get_quest_rewards;
+pub mod get_quest_reward;
 pub mod get_quest_stats;
 pub mod get_quest_updates;
 pub mod get_quests;
@@ -18,7 +18,7 @@ pub use create_quest::*;
 use dcl_crypto::Address;
 pub use delete_quest::*;
 pub use get_quest::*;
-pub use get_quest_rewards::*;
+pub use get_quest_reward::*;
 pub use get_quest_stats::*;
 pub use get_quest_updates::*;
 pub use get_quests::*;
